refactor(docs): extract source/manifest scan into lib/source-extraction.ts
Pure relocation of the hook/slot registration scanners (extractHookRegistrations,
extractSlotRegistrations) and the plugin-manifest readers (readPluginManifestDirectory,
readCorePluginManifests, readOfficialPluginManifests, readOfficialPluginCatalog) plus
the HookRegistrationDoc/PluginManifestDoc types out of scripts/docs/generate.ts into a
new leaf module. Function bodies byte-for-byte unchanged; generate.ts imports the four
consumed entry points back. docs:generate output verified byte-identical.
